Add files via upload
The final project in DS in Python. In our work we decided to use the data from the 2024, 2023 and 2022 F1 championships to predict which racers would finish on which places during the Grand Prix Emilia-Romani race, which took place on May 19. In total we collected 31 records from revious races and trained 4 Machine Learning models: Linear Regression, AutoRegressive Integrated Moving Average, Simple Exponential Smoothing and Holt’s Linear Smoothing with different hyperparameters. We also used the Miami Grand Prix race (the 31th record) to test our forecasting performance and based on the results chose the best model. After determining, that the Linear Regression model gave us best predictions, we once again trained the LR model to include the data from the last race from Miami and forecasted the possible top 5 racers in the Emilia-Romani race which could be seen at the very bottom. As the result, we correctly predicted 4 out of 5 top places!)